expect jdk code
