+++ fleeksoft-io/src/com/fleeksoft/io/Constants.kt
@@ -1,0 +1,9 @@
+package com.fleeksoft.io
+
+object Constants {
+    const val DEFAULT_CHAR_BUFFER_SIZE: Int = 8192
+    const val DEFAULT_BYTE_BUFFER_SIZE: Int = 8192
+
+    /** Maximum skip-buffer size  */
+    const val maxSkipBufferSize: Int = 8192
+}
