--- fleeksoft-io/src/com/fleeksoft/io/exception/IOException.kt
@@ -1,3 +1,0 @@
-package com.fleeksoft.io.exception
-
-class IOException(msg: String) : Exception(msg)
