+++ fleeksoft-io/src@jvmAndAndroid/com/fleeksoft/io/BufferedReader.kt
@@ -1,0 +1,3 @@
+package com.fleeksoft.io
+
+actual typealias BufferedReader = java.io.BufferedReader
