+++ fleeksoft-io/src@jvmAndAndroid/com/fleeksoft/io/InputStream.kt
@@ -1,0 +1,3 @@
+package com.fleeksoft.io
+
+actual typealias InputStream = java.io.InputStream
