+++ fleeksoft-io/src@jvmAndAndroid/com/fleeksoft/io/InputStreamReader.kt
@@ -1,0 +1,3 @@
+package com.fleeksoft.io
+
+actual typealias InputStreamReader = java.io.InputStreamReader
