+++ fleeksoft-io/src@jvmAndAndroid/com/fleeksoft/io/Reader.kt
@@ -1,0 +1,3 @@
+package com.fleeksoft.io
+
+actual typealias Reader = java.io.Reader
