+++ fleeksoft-io/src@jvmAndAndroid/com/fleeksoft/io/StringReader.kt
@@ -1,0 +1,3 @@
+package com.fleeksoft.io
+
+actual typealias StringReader = java.io.StringReader
