+++ fleeksoft-io/src@jvmAndAndroid/com/fleeksoft/io/exception/IOException.kt
@@ -1,0 +1,5 @@
+package com.fleeksoft.io.exception
+
+actual typealias IOException = java.io.IOException
+actual typealias VirtualMachineError = java.lang.VirtualMachineError
+actual typealias OutOfMemoryError = java.lang.OutOfMemoryError
