+++ fleeksoft-io/src@nonJvm/com/fleeksoft/io/InputStreamReader.kt
@@ -1,0 +1,16 @@
+package com.fleeksoft.io
+
+actual class InputStreamReader: Reader() {
+    override fun read(cbuf: CharArray, offset: Int, length: Int): Int {
+        TODO("Not yet implemented")
+    }
+
+    override fun ready(): Boolean {
+        TODO("Not yet implemented")
+    }
+
+    override fun close() {
+        TODO("Not yet implemented")
+    }
+
+}
