+++ fleeksoft-io/src@nonJvm/com/fleeksoft/io/Reader.kt
@@ -1,10 +1,26 @@
 package com.fleeksoft.io
 
-expect abstract class Reader {
+import com.fleeksoft.io.exception.IOException
+import kotlin.math.min
 
-    open fun read(): Int
+actual abstract class Reader {
 
-    abstract fun read(cbuf: CharArray, offset: Int, length: Int): Int
+    actual open fun read(): Int {
+        val cb = CharArray(1)
+        return if (read(cb, 0, 1) == -1) -1
+        else cb[0].code
+    }
+
+    actual abstract fun read(cbuf: CharArray, offset: Int, length: Int): Int
+
+    fun readString(length: Int): String {
+        val cbuf = CharArray(length)
+        read(cbuf)
+        return cbuf.concatToString()
+    }
+
+    /** Skip buffer, null until allocated  */
+    private var skipBuffer: CharArray? = null
 
 
     /**
@@ -20,7 +36,22 @@ expect abstract class Reader {
      * @throws     IllegalArgumentException  If `n` is negative.
      * @throws     IOException  If an I/O error occurs
      */
-    open fun skip(n: Long): Long
+    actual open fun skip(n: Long): Long {
+        require(n >= 0L) { "skip value is negative" }
+        return implSkip(n)
+    }
+
+    private fun implSkip(n: Long): Long {
+        val nn = min(n.toInt(), Constants.maxSkipBufferSize)
+        if ((skipBuffer == null) || (skipBuffer!!.size < nn)) skipBuffer = CharArray(nn)
+        var r = n
+        while (r > 0) {
+            val nc = read(skipBuffer!!, 0, min(r.toDouble(), nn.toDouble()).toInt())
+            if (nc == -1) break
+            r -= nc.toLong()
+        }
+        return n - r
+    }
 
     /**
      * Tells whether this stream supports the mark() operation. The default
@@ -29,7 +60,9 @@ expect abstract class Reader {
      *
      * @return true if and only if this stream supports the mark operation.
      */
-    open fun markSupported(): Boolean
+    actual open fun markSupported(): Boolean {
+        return false
+    }
 
 
     /**
@@ -41,7 +74,7 @@ expect abstract class Reader {
      *
      * @throws     IOException  If an I/O error occurs
      */
-    open fun ready(): Boolean
+    actual abstract fun ready(): Boolean
 
     /**
      * Marks the present position in the stream.  Subsequent calls to reset()
@@ -56,7 +89,9 @@ expect abstract class Reader {
      * @throws     IOException  If the stream does not support mark(),
      * or if some other I/O error occurs
      */
-    open fun mark(readAheadLimit: Int)
+    actual open fun mark(readAheadLimit: Int) {
+        throw IOException("mark() not supported")
+    }
 
     /**
      * Resets the stream.  If the stream has been marked, then attempt to
@@ -72,7 +107,9 @@ expect abstract class Reader {
      * or if some other I/O error occurs
      */
 
-    open fun reset()
+    actual open fun reset() {
+        throw IOException("reset() not supported")
+    }
 
     /**
      * Closes the stream and releases any system resources associated with
@@ -82,5 +119,5 @@ expect abstract class Reader {
      *
      * @throws     IOException  If an I/O error occurs
      */
-    abstract fun close()
+    actual abstract fun close()
 }
