+++ fleeksoft-io/src@nonJvm/com/fleeksoft/io/StreamDecoder.kt
@@ -18,7 +18,7 @@ class StreamDecoder(source: InputStream, charset: Charset = UTF_8.INSTANCE) : Re
         .onMalformedInput(CodingErrorAction.REPLACE)
         .onUnmappableCharacter(CodingErrorAction.REPLACE)
 
-    private var bb: ByteBuffer = ByteBuffer.allocate(DEFAULT_BYTE_BUFFER_SIZE);
+    private var bb: ByteBuffer = ByteBuffer.allocate(Constants.DEFAULT_BYTE_BUFFER_SIZE);
 
     // Exactly one of these is non-null
     private var source: InputStream? = source
@@ -130,7 +130,7 @@ class StreamDecoder(source: InputStream, charset: Charset = UTF_8.INSTANCE) : Re
         }
     }
 
-    fun implRead(cbuf: CharArray, off: Int, end: Int): Int {
+    private fun implRead(cbuf: CharArray, off: Int, end: Int): Int {
         // In order to handle surrogate pairs, this method requires that
         // the invoker attempt to read at least two characters.  Saving the
         // extra character, if any, at a higher level is easier than trying
@@ -215,7 +215,7 @@ class StreamDecoder(source: InputStream, charset: Charset = UTF_8.INSTANCE) : Re
 
     private fun inReady(): Boolean {
         return try {
-            ((this.source != null) && (this.source!!.exhausted().not())) // ## RBC.available()?
+            ((this.source != null) && (this.source!!.available() > 0)) // ## RBC.available()?
         } catch (x: IOException) {
             false
         }
