+++ fleeksoft-io/src@nonJvm/com/fleeksoft/io/internal/Assert.kt
