+++ fleeksoft-io/src@nonJvm/com/fleeksoft/io/internal/ObjHelper.kt
