+++ kotlinx-io/src/com/fleeksoft/io/SourceExt.kt
@@ -2,5 +2,4 @@ package com.fleeksoft.io
 
 import kotlinx.io.Source
 
-
-fun InputStream.Companion.from(source: Source): InputStream = InputStreamKotlinx(source)
+fun Source.toInputStream(): InputStream = InputStreamKotlinx(this)
